pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app", "app.openapi", "config"]

[tool.setuptools.package-data]
app = ["openapi/openapi.yaml"]

[project]
name = "falcon-url-checker"
version = "0.1.0"
description = "Falcon: проверка URL/DNS/TLS/HTTP и документация OpenAPI"
requires-python = ">=3.11"
dependencies = [
    "dnspython==2.8.0",
    "falcon==4.2.0",
    "gunicorn==25.3.0",
    "httpx==0.28.1",
    "pydantic==2.12.5",
    "pytest==9.0.3",
    "pyyaml==6.0.3",
    "python-dotenv==1.2.2",
    "ruff==0.15.9",
    "webtest==3.0.7",
]

[tool.pytest.ini_options]
pythonpath = ["."]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
